FSM detects a new lap and reacts accordingly
FSM is set to get ADV saved data. These data are fed into the algorithm for finding a new lap start by auto-correlation.
When it finds a start, the FSM transitions into a running state. In this state the FSM reads the saved ADC values and changes the car's speed accordingly by a hard-code thresholds for bends and straight sections.
After the saved-data-counter runs out, the FSM goes into a stopped state. It stays in this state for 10 seconds. Then rolls back into the reference lap state while clearing the buffers to ensure data readings from scratch.