add MAP function
Add MAP function: This function saves the track into stamped with distance and time. This is only the first version thus needs testing. Added a function to print the saved map over BLE DBG.

Add DEBUG state for FSM: Introduced a new state for FSM to simplify the process of debugging the FSM's behaviour.